requirements.txt
numpy>=1.21.0
scikit-learn>=1.0.0
openai>=1.0.0
sentence-transformers>=2.2.0
groq>=0.3.0
python-dotenv>=0.19.0